% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadInterface}
\alias{loadInterface}
\title{Launches the RHRV graphical interface opening a web browser}
\usage{
loadInterface(port = 1971L)
}
\arguments{
\item{port}{the port where to listen; 1971 by default.}
}
\description{
Launches the RHRV graphical interface opening a web browser
}
\author{
Patricia Guillin
}

